# api_specs.yaml
openapi: 3.0.0
info:
  title: Flash Learn API
  description: API used by Flash learn system
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
    description: Example server URL
paths:
...
  /deck/{id}/card/total:
    get:
      summary: Get total number of cards in deck with id
      operationId: getTotalCards
      responses:
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Total number of cards in deck
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 42
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "An error occurred"
        code:
          type: integer
          example: 500